+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 <h1 align="center">Changelogs</h1>
 
+## :bookmark: [React-Auctions v0.4.0](https://github.com/tklein1801/React-Auctions/releases/tag/v0.4.0)
+
+- Profile
+  - Create a new offer
+  - Show created offers
+  - Get status of created offers
+- Minor style & component updates
+
 ## :bookmark: [React-Auctions v0.3.0](https://github.com/tklein1801/React-Auctions/releases/tag/v0.3.0)
 
 - Add usernames using Firestore